+++ test/runTests.html
@@ -24,7 +24,7 @@
     <script src="../js/hicResolutionSelector.js"></script>
     <script src="../js/hicBrowser.js"></script>
     <script src="../js/contactMatrixView.js"></script>
-    <script src="../js/hicRuler.js"></script>
+    <script src="../js/yRuler.js"></script>
     <script src="../js/scrollbarWidget.js"></script>
     <script src="../js/sweepZoom.js"></script>
     <script src="../js/hicState.js"></script>
